fix: Address code review findings in security contracts
- Fix infinite recursion in VFIDEAccessControl.getRoleMemberCount()
- Add gas limit to CircuitBreaker emergency controller call
- Fix return type mismatch in VFIDETokenV2.getPriorVotes() (uint224)
- Add require() to virtual functions in WithdrawalQueue for safety
- Add gas limit to AdminMultiSig.executeProposal() low-level call

All code review issues resolved for production readiness.

+++ contracts/VFIDETokenV2.sol
@@ -238,7 +238,7 @@ contract VFIDETokenV2 is ERC20, VFIDEAccessControl, VFIDEReentrancyGuard {
     function getPriorVotes(address _account, uint256 _blockNumber) 
         external 
         view 
-        returns (uint96) 
+        returns (uint224) 
     {
         require(_blockNumber < block.number, "VFIDETokenV2: not yet determined");
